+++ threedi_modelchecker/model_checks.py
@@ -6,32 +6,35 @@ class ThreediModelChecker:
     def __init__(self, threedi_db):
         self.db = threedi_db
         self.schema = ModelSchema(self.db)
+        self.schema.validate_schema()
         self.config = Config(self.models)
 
     @property
     def models(self):
+        """Returns a list of declared models"""
         return self.schema.declared_models
 
-    def get_model_errors(self):
-        model_errors = []
-        session = self.db.get_session()
-        for check in self.config.checks:
-            model_errors += check.get_invalid(session)
-        return model_errors
+    def errors(self):
+        """Iterates and applies checks, returning any failing rows.
 
-    def get_model_error_iterator(self):
+        :return: Tuple of the applied check and the failing row.
+        """
         session = self.db.get_session()
-        for check in self.config.checks:
+        for check in self.checks():
             model_errors = check.get_invalid(session)
             for error_row in model_errors:
                 yield check, error_row
+
+    def checks(self):
+        """Iterates over all configured checks
+
+        :return: implementations of BaseChecks
+        """
+        for check in self.config.checks:
+            yield check
 
     def check_table(self, table):
         pass
 
     def check_column(self, column):
         pass
-
-    def apply(self, check):
-        """Applies the check and returns any invalid rows"""
-        return check.get_invalid(self.db.get_session())